eva_worker/requirements.txt
psycopg2-binary
openai
requests
